reorganized the massive color file into a folder of several smaller files
