+++ shoes-core/lib/shoes/color/system_color.rb
@@ -1,0 +1,6 @@
+class Shoes
+  class Color
+    class SystemColor
+    end
+  end
+end
